Exception handling in Location Service completed.

+++ pluq-meter-report/src/main/java/com/meterreport/pluqmeterreport/services/LocationService.java
@@ -1,5 +1,6 @@
 package com.meterreport.pluqmeterreport.services;
 
+import com.meterreport.pluqmeterreport.errors.customErrors.LocationAlreadyExistsException;
 import com.meterreport.pluqmeterreport.errors.customErrors.LocationNotFoundException;
 import com.meterreport.pluqmeterreport.models.location.Location;
 import com.meterreport.pluqmeterreport.repos.LocationRepository;
@@ -28,7 +29,7 @@ public class LocationService {
         return location;
     }
 
-    public List<Location> getAllLocations() {
+    public List<Location> getAllLocations() throws LocationNotFoundException {
         List<Location> locationList = locationRepository.findAll();
 
         if (locationList.isEmpty()) {
@@ -38,19 +39,30 @@ public class LocationService {
         return locationList;
     }
 
-    public Location saveLocation(Location location) {
+    public Location saveLocation(Location location) throws LocationAlreadyExistsException {
+        if(locationRepository.findById(location.getId()).isPresent()) {
+            throw new LocationAlreadyExistsException("Location already exists with id: " + location.getId());
+        }
+
         return locationRepository.save(location);
     }
 
+//    TODO: Finish exception handling for this method.
     public List<Location> saveLocationsList(List<Location> locationsList) {
         return locationRepository.saveAll(locationsList);
     }
 
-    public void delete(String locationId) {
+    public void delete(String locationId) throws LocationNotFoundException {
+        if(locationRepository.findById(locationId).isEmpty()) {
+            throw new LocationNotFoundException("Could not find location by id: " + locationId);
+        }
         locationRepository.deleteById(locationId);
     }
 
-    public void deleteAllLocations() {
+    public void deleteAllLocations() throws LocationNotFoundException {
+        if (locationRepository.findAll().isEmpty()) {
+            throw new LocationNotFoundException("Could not find any saved locations.");
+        }
         locationRepository.deleteAll();
     }
 }
